Added automatic language version scraping (#249)
* added language versioning

Added a run that will run a python program to store all the version of each of the languages in a json file. The output will then be read to populate the version field of Languages. Additionallly, added a test for adding the version numbers.

* Update dev-server.js

* Made some minor changes to previously modified files.

Cleared out commented out debugging statements, and created a temporary contest in which the version run is ran.